LayoutTests/animations/many-pseudo-animations-expected.txt

Tests that animations on pseudo-elements do not eventually stop working.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS animationsStartedThisFrame is numElements
PASS animationsStartedThisFrame is numElements
PASS animationsStartedThisFrame is numElements
PASS animationsStartedThisFrame is numElements
PASS animationsStartedThisFrame is numElements
PASS animationsStartedThisFrame is numElements
PASS animationsStartedThisFrame is numElements
PASS animationsStartedThisFrame is numElements
PASS animationsStartedThisFrame is numElements
PASS animationsStartedThisFrame is numElements
PASS successfullyParsed is true

TEST COMPLETE
